Issue-422: Moved UIMovie and UITv to core module
