+++ app/src/main/java/com/kirchhoff/movies/repository/movie/IMovieRepository.kt
@@ -1,11 +1,11 @@
 package com.kirchhoff.movies.repository.movie
 
+import com.kirchhoff.movies.core.data.UIMovie
 import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
 import com.kirchhoff.movies.data.ui.details.movie.UIMovieDetails
 import com.kirchhoff.movies.data.ui.details.movie.UITrailersList
-import com.kirchhoff.movies.data.ui.main.UIMovie
 
 interface IMovieRepository {
     suspend fun fetchDetails(movieId: Int): Result<UIMovieDetails>
