+++ app/src/main/java/com/kirchhoff/movies/repository/movie/MovieRepository.kt
@@ -1,12 +1,12 @@
 package com.kirchhoff.movies.repository.movie
 
+import com.kirchhoff.movies.core.data.UIMovie
 import com.kirchhoff.movies.core.repository.BaseRepository
 import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
 import com.kirchhoff.movies.data.ui.details.movie.UIMovieDetails
 import com.kirchhoff.movies.data.ui.details.movie.UITrailersList
-import com.kirchhoff.movies.data.ui.main.UIMovie
 import com.kirchhoff.movies.mapper.discover.IDiscoverMapper
 import com.kirchhoff.movies.mapper.movie.IMovieDetailsMapper
 import com.kirchhoff.movies.network.services.MovieService
