+++ app/src/main/java/com/kirchhoff/movies/ui/screens/core/movies/adapter/MovieVH.kt
@@ -4,10 +4,10 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import com.kirchhoff.movies.R
+import com.kirchhoff.movies.core.data.UIMovie
 import com.kirchhoff.movies.core.extensions.downloadPoster
 import com.kirchhoff.movies.core.extensions.inflate
 import com.kirchhoff.movies.core.ui.recyclerview.adapter.viewholder.BaseVH
-import com.kirchhoff.movies.data.ui.main.UIMovie
 import com.kirchhoff.movies.keywordsview.RatingTextView
 
 class MovieVH(parent: ViewGroup) : BaseVH<UIMovie>(parent.inflate(R.layout.item_movie)) {
