+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/movies/MoviesFragment.kt
@@ -1,10 +1,10 @@
 package com.kirchhoff.movies.ui.screens.main.movies
 
 import com.kirchhoff.movies.R
+import com.kirchhoff.movies.core.data.UIMovie
 import com.kirchhoff.movies.core.ui.paginated.PaginatedScreenFragment
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.core.ui.recyclerview.adapter.BaseRecyclerViewAdapter
-import com.kirchhoff.movies.data.ui.main.UIMovie
 import com.kirchhoff.movies.ui.screens.core.movies.adapter.MoviesListAdapter
 import com.kirchhoff.movies.ui.screens.details.movie.MovieDetailsFragment
 import org.koin.android.viewmodel.ext.android.viewModel
