+++ app/src/main/java/com/kirchhoff/movies/ui/screens/similar/movie/SimilarMoviesFragment.kt
@@ -3,10 +3,10 @@ package com.kirchhoff.movies.ui.screens.similar.movie
 import android.os.Bundle
 import android.view.View
 import com.kirchhoff.movies.R
+import com.kirchhoff.movies.core.data.UIMovie
 import com.kirchhoff.movies.core.ui.paginated.PaginatedScreenFragment
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.core.ui.recyclerview.adapter.BaseRecyclerViewAdapter
-import com.kirchhoff.movies.data.ui.main.UIMovie
 import com.kirchhoff.movies.ui.screens.core.movies.adapter.MoviesListAdapter
 import com.kirchhoff.movies.ui.screens.details.movie.MovieDetailsFragment
 import org.koin.android.viewmodel.ext.android.viewModel
