+++ app/src/main/java/com/kirchhoff/movies/ui/screens/similar/tv/SimilarTvsFragment.kt
@@ -3,10 +3,10 @@ package com.kirchhoff.movies.ui.screens.similar.tv
 import android.os.Bundle
 import android.view.View
 import com.kirchhoff.movies.R
+import com.kirchhoff.movies.core.data.UITv
 import com.kirchhoff.movies.core.ui.paginated.PaginatedScreenFragment
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.core.ui.recyclerview.adapter.BaseRecyclerViewAdapter
-import com.kirchhoff.movies.data.ui.main.UITv
 import com.kirchhoff.movies.ui.screens.core.tvs.adapter.TvsListAdapter
 import com.kirchhoff.movies.ui.screens.details.tv.TvDetailsFragment
 import org.koin.android.viewmodel.ext.android.viewModel
