+++ core/src/main/java/com/kirchhoff/movies/core/data/UITv.kt
@@ -1,12 +1,12 @@
-package com.kirchhoff.movies.data.ui.main
+package com.kirchhoff.movies.core.data
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class UIMovie(
+data class UITv(
     val id: Int,
-    val title: String?,
+    val name: String?,
     val posterPath: String?,
     val backdropPath: String?,
     val voteAverage: Float?
